requirements.txt
#selenium==3.141.0
PyInstaller==3.5
pytest==5.3.1
# for not using selenium
beautifulsoup4>=4.10.0
requests
lxml
python-telegram-bot
SQLAlchemy